requirements.txt
pyyaml
requests
psycopg2-binary
python-dotenv
